move jep script engine classes into their own files
