+++ src/main/java/edu/ucsb/nceas/mdqengine/dispatch/Dispatcher.java
@@ -1,38 +1,21 @@
 package edu.ucsb.nceas.mdqengine.dispatch;
 
-import edu.ucsb.nceas.mdqengine.MDQconfig;
 import edu.ucsb.nceas.mdqengine.model.Output;
 import edu.ucsb.nceas.mdqengine.model.Result;
 import edu.ucsb.nceas.mdqengine.model.Status;
 
-import org.apache.commons.configuration2.ex.ConfigurationException;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
-import jep.SubInterpreter;
-import jep.MainInterpreter;
-import jep.JepException;
-
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.io.IOException;
-import java.io.Reader;
-
-import javax.script.Bindings;
 import javax.script.ScriptContext;
 import javax.script.ScriptEngine;
-import javax.script.ScriptEngineFactory;
 import javax.script.ScriptException;
-import javax.script.SimpleBindings;
 import javax.script.ScriptEngineManager;
 import javax.script.Invocable;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.Collections;
-import java.util.List;
 
 public class Dispatcher {
 
@@ -192,6 +175,16 @@ public class Dispatcher {
 		return env.equalsIgnoreCase(currentEnv);
 	}
 
+	public void close() {
+		if (engine instanceof JepScriptEngine) {
+			JepScriptEngine jepEngine = (JepScriptEngine) engine;
+			jepEngine.close();
+			log.debug("Closed Jep interpreter");
+		} else {
+			log.debug("Didn't close Jep interpreter");
+		}
+	}
+
 	protected Dispatcher() {
 	}
 
@@ -270,342 +263,4 @@ public class Dispatcher {
 	public void setBindings(Map<String, Object> bindings) {
 		this.bindings = bindings;
 	}
-
-	/**
-	 * Implements {@link javax.script.ScriptEngine}
-	 * 
-	 */
-	public class JepScriptEngine implements ScriptEngine {
-
-		private SubInterpreter jepInterpreter = null;
-		private Bindings bindings = new SimpleBindings();
-		private Bindings globalBindings = new SimpleBindings();
-		private ScriptEngineFactory factory = null;
-
-		/**
-		 * Create a JepScriptEngine
-		 * 
-		 * The JepScriptEngine initalizes a jep SharedInterpreter, and relevant
-		 * ScriptEngine methods are called on the interpreter. Only relevant
-		 * ScriptEngine methods are implemented, the rest will error if they
-		 * are called. The interpreter is configured to find the jep path listed
-		 * in metadig.properties as `jep.path` for production deployments. For testing
-		 * on other machines, the configured path can be set using the environment
-		 * variable JEP_LIBRARY_PATH.
-		 * 
-		 * @throws RuntimeException              for script or configuration errors
-		 * @throws UnsupportedOperationException for methods not implemented
-		 */
-		public JepScriptEngine() {
-			// first look for an env var (this is mostly for local testing)
-			String pythonFolder = System.getenv("JEP_LIBRARY_PATH");
-
-			// then look in mdq config
-			if (pythonFolder == null) {
-				try {
-					MDQconfig cfg = new MDQconfig();
-					pythonFolder = cfg.getString("jep.path");
-				} catch (ConfigurationException ce) {
-					throw new RuntimeException("Error reading metadig configuration, ConfigurationException: " + ce);
-				} catch (IOException io) {
-					throw new RuntimeException("Error reading metadig configuration, IOException: " + io);
-				}
-			}
-
-			// define the jep library path
-			String jepPath = pythonFolder + "/libjep.jnilib";
-
-			if (!Files.exists(Path.of(jepPath))) {
-				jepPath = pythonFolder + "/libjep.so";
-			}
-
-			try {
-				// set path for jep executing python
-				MainInterpreter.setJepLibraryPath(jepPath);
-
-			} catch (JepException e) {
-				throw new RuntimeException("Error setting configurating Jep interpreter: " + e);
-			}
-
-			try {
-				// create the interpreter for python executing
-				jepInterpreter = new SubInterpreter();
-			} catch (JepException e) {
-				throw new RuntimeException("Error initializing Jep interpreter: " + e);
-			}
-		}
-
-		/**
-		 * Mapping of eval method for ScriptEngine to the jep SharedInterpreter
-		 * 
-		 * Instead of eval we actually implement exec, which allows us to run an
-		 * arbitrary number of statments as opposed to just one. Note that exec returns
-		 * void. To retrieve values from the python execution, `getValue` is required.
-		 * 
-		 * Since this method is designed to fit in with dispatch code running other
-		 * engines that return values from eval, it cannot return null. Returning a
-		 * boolean can make it look like code was executed and results available when
-		 * they might not be, so a null-looking value is returned and handled in the
-		 * dispatch code.
-		 * 
-		 * @param script The python script as a string
-		 * 
-		 * @throws RuntimeException If an error occurs during script execution.
-		 * @see jep.SharedInterpreter#exec()
-		 * 
-		 */
-		@Override
-		public Object eval(String script) throws RuntimeException {
-			// Implement the evaluation logic using the jepInterpreter
-			try {
-				jepInterpreter.exec(script);
-				return "_NA_";
-			} catch (JepException e) {
-				throw new RuntimeException(e);
-			}
-		}
-
-		public void close() {
-			try {
-				jepInterpreter.close();
-			} catch (JepException e) {
-				throw new RuntimeException(e);
-			}
-		}
-
-		/**
-		 * Retrieves a value associated with the specified key
-		 *
-		 * @param key The key representing the value to retrieve.
-		 * @return The value associated with the key, if found.
-		 * @throws RuntimeException If an error occurs during value retrieval.
-		 * @see jep.SharedInterpreter#getValue(String)
-		 */
-		@Override
-		public Object get(String key) {
-			try {
-				Object object = jepInterpreter.getValue(key);
-				return object;
-			} catch (JepException e) {
-				throw new RuntimeException(e);
-			}
-		}
-
-		/**
-		 * Set a value associated with the specified key
-		 *
-		 * @param key   The key to set.
-		 * @param value The value to assign to the key.
-		 * @throws RuntimeException If an error occurs.
-		 * @see jep.SharedInterpreter#set(String, Object)
-		 */
-		@Override
-		public void put(String key, Object value) {
-			try {
-				jepInterpreter.set(key, value);
-			} catch (JepException e) {
-				throw new RuntimeException(e);
-			}
-		}
-
-		/**
-		 * Set the script engine factory
-		 *
-		 * @param fact The script engine factory to assign
-		 */
-		protected void setFactory(ScriptEngineFactory fact) {
-			this.factory = fact;
-		}
-
-		/**
-		 * Get the script engine factory
-		 *
-		 */
-		@Override
-		public ScriptEngineFactory getFactory() {
-			if (this.factory == null)
-				this.factory = new JepScriptEngineFactory();
-			return this.factory;
-		}
-
-		/**
-		 * Get bindings
-		 * 
-		 * Not sure exactly how/if this works, got this from an old Jep release
-		 * 
-		 * @param scope
-		 * @return a Bindings value
-		 */
-		@Override
-		public Bindings getBindings(int scope) {
-			if (scope == ScriptContext.ENGINE_SCOPE) {
-				return this.bindings;
-			}
-
-			return this.globalBindings;
-		}
-
-		public String getEngineName() {
-			return "python";
-		}
-
-		@Override
-		public void setBindings(Bindings bindings, int scope) {
-			throw new UnsupportedOperationException(
-					"setBindings(Bindings, int) not supported for the JepScriptEngine class");
-		}
-
-		@Override
-		public Bindings createBindings() {
-			throw new UnsupportedOperationException("createBindings() not supported for the JepScriptEngine class");
-		}
-
-		@Override
-		public ScriptContext getContext() {
-			throw new UnsupportedOperationException("getContext() not supported for the JepScriptEngine class");
-		}
-
-		@Override
-		public void setContext(ScriptContext context) {
-			throw new UnsupportedOperationException(
-					"setContext(ScriptContext) is not implemented for the JepScriptEngine class");
-		}
-
-		public Object eval(Reader reader, Bindings bindings) throws ScriptException {
-			throw new UnsupportedOperationException(
-					"eval(Reader, Bindings) is not implemented for the JepScriptEngine class");
-		}
-
-		@Override
-		public Object eval(String script, Bindings bindings) throws ScriptException {
-			throw new UnsupportedOperationException(
-					"eval(String, Bindings) is not implemented for the JepScriptEngine class");
-		}
-
-		@Override
-		public Object eval(Reader reader) throws ScriptException {
-			throw new UnsupportedOperationException("eval(Reader) is not implemented for the JepScriptEngine class");
-		}
-
-		@Override
-		public Object eval(Reader reader, ScriptContext context) throws ScriptException {
-			throw new UnsupportedOperationException(
-					"eval(Reader, ScriptContext) is not implemented for the JepScriptEngine class");
-		}
-
-		@Override
-		public Object eval(String script, ScriptContext context) throws ScriptException {
-			throw new UnsupportedOperationException(
-					"eval(String, ScriptContext) is not implemented for the JepScriptEngine class");
-		}
-
-	}
-
-	public class JepScriptEngineFactory implements ScriptEngineFactory {
-
-		private static List<String> names;
-
-		private static List<String> extensions;
-
-		private static List<String> mimeTypes;
-
-		static {
-			names = new ArrayList<>(1);
-			names.add("jep");
-			names = Collections.unmodifiableList(names);
-
-			extensions = new ArrayList<>(1);
-			extensions.add("py");
-			extensions = Collections.unmodifiableList(extensions);
-
-			mimeTypes = new ArrayList<>(0);
-			mimeTypes = Collections.unmodifiableList(mimeTypes);
-		}
-
-		@Override
-		public String getEngineName() {
-			return "jep";
-		}
-
-		@Override
-		public String getEngineVersion() {
-			return "2.x";
-		}
-
-		@Override
-		public List<String> getExtensions() {
-			return extensions;
-		}
-
-		@Override
-		public String getLanguageName() {
-			return "python";
-		}
-
-		@Override
-		public String getLanguageVersion() {
-			throw new UnsupportedOperationException(
-					"getLanguageVersion is not implemented for the JepScriptEngineFactory class");
-		}
-
-		@Override
-		public String getMethodCallSyntax(String obj, String method, String... args) {
-			throw new UnsupportedOperationException("getMethodCallSyntax is not implemented");
-		}
-
-		@Override
-		public List<String> getMimeTypes() {
-			return mimeTypes;
-		}
-
-		@Override
-		public List<String> getNames() {
-			return names;
-		}
-
-		@Override
-		public String getOutputStatement(String o) {
-			throw new UnsupportedOperationException("getOutputStatement is not implemented");
-		}
-
-		@Override
-		public Object getParameter(String p) {
-			if (p == null)
-				return null;
-
-			if (p.equals(ScriptEngine.ENGINE))
-				return getEngineName();
-
-			if (p.equals(ScriptEngine.ENGINE_VERSION))
-				return getEngineVersion();
-
-			if (p.equals(ScriptEngine.NAME))
-				return "jep";
-
-			if (p.equals(ScriptEngine.LANGUAGE))
-				return getLanguageName();
-
-			if (p.equals(ScriptEngine.LANGUAGE_VERSION))
-				return getLanguageVersion();
-
-			return null;
-		}
-
-		@Override
-		public String getProgram(String... statements) {
-			throw new UnsupportedOperationException(
-					"getProgram is not implemented for the JepScriptEngineFactory class");
-		}
-
-		@Override
-		public ScriptEngine getScriptEngine() {
-
-			JepScriptEngine e = new JepScriptEngine();
-			e.setFactory(this);
-			return e;
-
-		}
-
-	}
-
 }
